Begin methods to get start and possible moves
Complete a method that gets the starting position of the knight from
the user.

Work on method to return all possible moves of a knight from that
knights position. Finish handling moves that return negative array
values off board so they are not counted as possible moves. Continue
working on handling positive array values that end up off the board.